requirements.txt
fastapi[standard]
sqlmodel>=0.0.37
sqlalchemy>=2.0.14,<2.1.0
pydantic>=2.0
pyjwt>=2.0
pwdlib[argon2]
python-dotenv
apscheduler>=3.10.0
httpx>=0.23.0